Fix dblatex-3.0 image path problem
dblatex-3.0 uses the XSL parameter "keep.relative.image.uris" (with
default value "0") which resolves image URIs against xml:base. That is,
image paths 'fileref="images/..."' are preceded with the source file's
directory name. That's not what we need, so we have to set this
parameter to "1" to leave fileref attributes unchanged.