[make] Remove configure option "--with-xsl"
Now users cannot configure gimp-help-2 to use a custom path to their
DocBook XSL stylesheets any more. Instead, they are encouraged (ok,
they are forced...) to configure their XML catalog so that the base
canonical URI ("http://docbook.sourceforge.net/...") is resolved
automatically.
As a compensation, a simple test for working DocBook XSL stylesheets
is added which displays a warning if the canonical URI can't be
resolved and a hint how to fix that.