Removed too aggressive rule.
The affected line was modified in 2008. I think Konqueror is most likely
better with dealing with affected tags by now. This actually fixes the
missing link to our Bugzilla page, which was there, but simply not
displayed due to the CSS rule.
Kudos to Marco Ciampa for pointing this out.