Update Preferences/Help.
If the internal GIMP help browser is not installed, then the next 2.8
release will hide the listbox 'Help Browser to use' and always open the
help files in the system's standard web browser.
Remove obsolete description 'Web browser to use'. This option was
removed from GIMP already in 2009.
Remove obsolete line 'Linux screenshot' because it doesn't provide
useful information in this context and let's assume that this feature
only works on Linux, which is not the case.