autogen: honor the environment variable NOCONFIGURE.
This is a very common environment variable, nearly a standard since it
is cited in the "Build API" (followed by Flatpak for instance). GIMP
codebase have already honored this variable for some time now.
There are many cases when you don't want to run configure, in particular
when you just want to generate the configure file for a tarball, or
again when you want to do a VPATH build.