configure: use tar-pax tar format for make dist.
By default, automake uses tar-v7 which has a limitation of 99 characters
to paths. This results in many files not being saved inside the tarball,
while diplaying such error on a `make dist`:
> tar: gimp-help-2.9.5/some/long/path: file name is too long (max 99); not dumped
The tar-pax format is newer, without length limitation. See:
https://www.gnu.org/software/automake/manual/html_node/List-of-Automake-options.html
It is available since automake 1.9 but it turns out that our minimum
requirement is already automake 1.10 since commit da28313e but this was
only visible in the autogen.sh. Let's make the configure script
consistent as well.