Remove "Fade" page in the Edit menu.
Fade action does not exist anymore in GIMP 2.10 for various reasons. We
should not keep the page in our docs, this is misleading.

Note that the screenshot of the Edit menu in gimp-edit-menu should also
be redone as it still shows Fade in the list.

See: https://gitlab.gnome.org/GNOME/gimp/issues/3558