WIP commit: Try to use meson as build system.
It seems that meson isn't that well suited for
something that has different source and build
directory structures.

Maybe we need something similar to the yelp
module for meson.

Currently we can build pot files, xml files and
html files without xref and manually have to
move css, images and stylesheets for the
xml to html translations.

Also all files are always built.