tools: add --base option in xml2po to set base path from build to source
When using separate build and source directories the generated pot and po
files got comments about source po files including the redirection path
from build to source directory (e.g. ../../gimp-help).

This can cause a lot of unnecessary changes when committing by someone that
uses a different build directory path.

To fix this we added an option --base to set the base path to get from
the build directory to the source directory. This part of the filepath
is then removed from the filepaths set in po and pot comments.