po: fix sl/menus/colors/components.po
The Slovenian components.po hasn't been updated in a long time and could
cause make to fail because of a difference between msgid and msgstr with
regards to starting with a newline.

These two occurrences were part of script-fu code. I changed them so msgstr
is the same as msgid.

Besides that the remainder of that po file has been updated to the most
recent version of the components.pot file too.