configure: fix absent top_srcdir variable.
With double quotes, $(top_srcdir) would look like a shell command to
run, which is obviously wrong. Instead, switching to single quotes, this
is not interpreted in the configure script, hence can be replaced by the
corresponding automake's variable value at the next step.

This fixes the configure error:

> gimp-help/configure: line 2992: top_srcdir: command not found

Note that this error was breaking neither the configure step nor the
build. It was just displayed during configure.