tools: fix crash with older version of libxml2
The DamnedLies translation website relies on an older version
of libxml2 which apparently doesn't have an interator for
XmlAttr which means our pot files do not get regenerated
and our translations get stuck.

Let's work around this by using an alternative loop that
doesn't use an iterator.

Hopefully this is the only problem, but we will find out
after we commit this.