configure, m4macros: check python version and libxml2 availability.
We have this issue on Damned Lies (GNOME infra for translators) where
they run Python 3.9 while libxml2 module is only available for Python
3.6. We don't want to just enforce Python 3.6 because on contributors
machines, it may be otherwise (on my Fedora, libxml2 Python module is
for Python 3.9). So we could break local builds.
Also Python 3.6 is EOL in December, it does not seem like a wise choice
to make it a hard dependency for everyone just because the Damned Lies
container use old software.

Instead make a Python test at configure time, which would check not only
for the Python version, but also for the availability of the libxml2
module. Hence it would select the highest Python 3 which can import
libxml2. Then we would run our xml2po script explicitly with the Python
found at configure time (not relying on the shebang).

The implementation details are:

- Set Python 3.6.0 as our minimum requirement, like in GIMP. This is
  actually the lower Python 3 verson which is still supported (until end
  of year). No need to support older Python.
- Test first python3 and python (default system python), then python3.9,
  python3.8 down to 3.6.
- The m4 macro AM_PATH_PYTHON3() is taken from the GIMP repo, improved
  with the additional 'MODULE' argument to specify a module requirement.
  AM_PATH_PYTHON3() was itself a fork of the generic automake macro
  AM_PATH_PYTHON() (which would also look for Python 2 but we don't want
  this anymore obviously).
- This version of the AM_PATH_PYTHON3() macro can only handle a single
  Python module. If we needed at some point to have several module
  requirements, we should improve it a bit more.