docs: fix #256 Section 9 Varying brush size - misses one method
Actually both the keyboard and mouse shortcuts were missing.
Besides that I made some other changes (replacing varying by changing).

Now that we have added the shortcuts for both keyboard and mouse I feel
that the instructions for adding both kinds of shortcuts are a little
to extensive, but will leave them for now.
We can probably replace this by links in the manual that explain how to
change these shortcuts.