Reworked and extended getting (un)stuck section
Reduced the depth of descriptions by one level. This way each separate
getting stuck chapter is shown in our index, making it easier to find.
This also causes each chapter to be on its own page, which could also
help in making it easier to find what you are looking for.

Reworked the existing getting stuck chapters and added new ones. Most of
it is based on the following issues:
- Issue #297 Explain certain situations when painting will have no effect.
- Issue #296 Add to getting stuck a section about unhiding tool windows.
- Issue #295 Add to getting stuck a section about unhiding the Tool
Options Dialog.
- Issue #268 floating selection suggestions for stuck novice.
- Issue #140 Another point to getting unstuck (Layer group selected).