Issue #299 Remove unmaintained  ODF support
We have optional support for converting our help from docbook to
OpenDocument format (ODF).
For this support we need package docbook2odf which has been
unmaintained for a long time and was removed from Debian in 2019.

For this reason we are removing ODF support for now. If a well
maintained package appears and someone comes along to support it
for gimp-help, then we can reconsider including ODF support again.