build: improve language filtering in Quickreference Makefile
Quickreference has a translation for pl that is not yet in our general
translations. This pl was always filtered out.
We also always ignored the globally set LINGUAS.

Changed this to always limit building to languages set in LINGUAS if that
variable is set.
If it is not set then use QUICKREFERENCE_ALL_LINGUAS.

(cherry picked from commit 2f9e9ed06b591930e4c7128d2df82bb189682f66)