Revert "web: update the copyright dates."
This reverts commit 03e277abdf884c879338692da535d151ec41a60f.

Oups I had not realized that Jacob had already taken care of it in the
gimp-help-2-10 branch in a much more elegant and future-proof way. Also
it seems that testing.docs.gimp.org is meant to follow this 2.10 branch,
not the `master` branch anyway.

Reverting to avoid merge conflicts later on. I'll let Jacob cherry-pick
the right commit later.