Synchronize all po files for all languages
make po is not picking up old changes, possibly because usually po files
are only updated when a new translation is committed and some languages
have not been updated in a long time. Not having up to date po files makes
our translation percentage on our website incorrect too.

To fix that for now, let's synchronize them. We do that by touch-ing all
xml files in the /src folder.