build: bring back support for using GIMP's po files as Compendium
We had support for using GIMP's po files as Compendium that can be used
to add translations. This got removed in 2012 in commit 140bc842 when
the Makefiles got revised.

Since the help about GIMP_PO_ROOT never got removed nor the part that
checks GIMP's folders for po files, I think this was done by mistake, so
let's get support for this back.

If you set GIMP_PO_ROOT to GIMP's folder and touch all xml files in the
/src folder before make po, you will see that it improves the translation
percentage for several languages, especially the ones that are not very
complete. In some cases it increases the percentage by almost 10%.