authors: fix docbook output, personname requires a child element
We were getting make validate errors because we omitted a child element
inside <personname>.

Since splitting in firstname and surname is problematic with the authors
list we have, let's just use othername.

The main thing is that we separate the name from the list of languages to
be able to not translate the names of authors.