build: add US English isl file for the installer.
The original file seems to be lost through times, as well as any
translations. Fortunately Jernej could extract the US strings from a
compiled installer. So let's restart from here.

As for the other languages, it's only 5 strings anyway, so let's not
waste too much time and start on a clean slate.