build: scripts in their current form.
Committer's (Jehan) comment: I am pushing the script in their current
form as provided by Jernej, as a start point. These are not generic yet
and won't work for the CI so I will update them. This way, we will also
have the original version (which we know used to work in older releases)
to inspect in cases some of my changes end up wrong.

Note that I don't push the translation (.isl) files because the ones
kept by Jernej seem to be very old GIMP setup language files which don't
correspond at all to the messages in the gimp-help installer. We are not
sure where the good files went.