build: only propose the installed language as installer lang.
Unlike GIMP which comes with all localizations, we have one installer
per manual's available language. So we should assume that if you
specifically targetted a given manual installer, it means you understand
associated lang. I think this was the original logic in this file,
except that using the macro in an `#if FileExists()` structure was not
working. Also anyway some langs are inside Unofficial/ and other not.
Moreover not all file names end in .isl (some can end in .islu when you
look in InnoSetup repository, even though none of the current manual
lang are in this case… but someday, we could have one!). So let's just
specialize per language code.