Issue #317: Change categories of AUTHORS
This commit implements a simplification and more consistent use of
authors.xml.
It also makes the output in AUTHORS and preface.html more alike.

We don't use the creator tag anymore, but for now I didn't completely
remove it from the dtd, jsut made it not required and not used in our
current xsl files.

I added an extra contributor role translator and moved all contributors
with non English as language to that role.
All English creators and contributors now are contributors with the
documenter role.