Issue #282:  mentioning Image menu is confusing
Using image menu for our main menu is confusing, since Image is also
a part of the main menu. Just call it menu in this case.

Second issue is that apparently we had a different default mouse
cursor for the paths tool.
We correct that here and also add a link to the prefs where that
cursor can be changed.