gitlab-ci: build the Windows installer's .isl files in the www step.
I am still unsure where exactly it happens, but the LINGUAS env variable
does affect the help.*.isl creation. I.e. that instead of using the po
files, intltool-merge uses po-windows-installer/.intltool-merge-cache as
set in -c option, which won't contain contents of po files from
languages not in LINGUAS (the problem is that I'm not sure where this
gets filled to be affected by LINGUAS).

I could just remove the `-c` caching option, but on the other hand, it
feels to me like this behavior makes sense. So instead, I'm going to
build the isl files in the www-debian step, with no LINGUAS variable
(i.e. for all langs).