gitlab-ci, configure, build: new --enable-windows-installer option.
I was initially unsure if this option (same as in GIMP repo) was
necessary. But come to think of it, it's probably better not to force
building the .isl files as in most cases, especially as it requires
tools (intltool) not necessary in other parts of the build.

I also make the 2 other intltool check (for the `make update-po` call
inside po-windows-intaller/) not breaking the rest of the
internationalization. These are also not mandatory for most of the
localization process.