web: ad Open Graph metadata (sort of).
It's not exactly proper og tags as it should be 'property' attribute
(not 'name'). Yet reading the dockbook xsl files for websites, it looks
like they understand only the 'name' attribute and have no way to
transform it into property. Maybe we should look on making our own xsl
file to support this.

Anyway, looking this up a bit, it looks like some platforms where these
metadata are used may understand when using 'name' attribute. So let's
try as-is, as a first step.