build, po-windows-installer: add ability to leave translator comments.
intltool-extract does not consider the `;` isl-syntax comments because
we tell it it is in the ini format. So let's have our source in actual
ini format (POTFILES does not like to look for translated strings in
constructed files, only from source files), then create the .isl files
from it in 3 steps (first, transform the comments in isl format; then
merge all language in a single .isl file; last generate one .isl per
language).

Finally add some comments to explain what %1 and %2 are replaced with
(this was asked by Piotr Drąg on the gnome-i18n mailing list).