fix GIMP user folder paths in customize-splashscreen
In issue #329 it was noticed that the macOS user folder for GIMP was not
visible, even though it was supposed to be fixed in #171.
The invisibility was caused by an incorrect role "tex" instead of "html".

Besides that I noticed that for Windows we still showed the old XP path.
Since we don't support XP anymore, let's just replace it by the path in
more recent versions of Windows.

(cherry picked from commit 54a7a95ee9255aa363ce43fc128cc272cca4f31c)