Add several missing path help id's
Added id's for:
- gimp-path-merge-visible
- gimp-path-visible
- gimp-path-linked
- gimp-path-fill (in edit menu, similar to gimp-path-stroke, was mistakenly
called gimp-edit-path-fill)
- gimp-path-lock-strokes
- gimp-path-lock-position
- gimp-path-color-tag
- gimp-path-export

This way these topics can be found when searching for it in GIMP with F1.

Added description for Fill Path, changed menu entry from Path Tool to
Edit Path.

Updated the Paths dialog and Paths context menu images. Decided not to add
an image for the "Lock" icons at the top of the dialog, since they will
be gone in 3.0.