Improve Quickreference SVG and PDF output
On languages like Polish part of the title was overlapping Wilber and
there were also two lines in the first column cut off.

This is an extension of what was suggested in MR !69 by Anders Jonsson.

The Bitstream Vera Sans font is also less portable, so we default to
DejaVu Sans and define a few substitues because Pango was giving out
warnings for me on Windows.

Two make overlapping or cut-off text less likely we:
- Reduce the title font size a little bit
- Reduce the action font size a little bit
- Reduce the size of the left margin
- Enlarge the width of the columns

Also update the inkscape version in the metadata so Inkscape won't
complain about needing conversion when opening our svg.