Fix #232: Newsprint documentation referenced the old plugin.
The Newsprint filter was changed to a GEGL filter since 2.10, but the
documentation was never updated and there were considerable differences.
This commit updates the documentation to reflect the current plug-in,
including updated images and one removed image that seemed to be specific
to the old plugin.

(cherry picked from commit d239e1aac4cdafd082411aeb02e4abe09f573a90)