Add missing gimp-image-grid help id
The GIMP help id for configuring the grid is gimp-image-grid, so adding
that id to the documentation.

I choose not to remove the id's that include configure, since I think that
is a more descriptive name, that we maybe should use in GIMP too.

(cherry picked from commit 6ef05e0e910923ab7f365fb4228cddbdb6ae9b81)