Add gimp-file-save-as-template help id to Create Template
Even though gimp updated the name of the define for Create Template,
it did not change the id string. In gimp-help this string was updated,
causing help not found when asking for help for Create Template.

For now, just add back the old help id as an anchor.

(cherry picked from commit 58ba9c3d1ee67039635cba876da734d0191112cb)