Fix link in tone-mapping-tutorial.xml
The restore-micro-contrast link id does not have a descriptive text.
Changing to a "normal" link instead of a xref so that we can decide
in context what the link text is.

While we are at it, we also remove a TAB and spaces at line endings for
this paragraph.

(cherry picked from commit 348f693299f6fa8369587664f3b569a3988e0b67)