Add Keyboard Shortcuts Dialog documentation
Added a new Keyboard Shortcuts Dialog page in the Misc Dialogs section.
A shorter explanation was available in Enrich my GIMP: creating shortcuts
to menu commands, but I think this documentation fits better in the
Dialogs category.

Updated the help id to point to this page instead of the menu entry.
The menu entry documentation now also points to this page.

The page on creating shortcuts to menu functions was revised to point to
this page for the dialog documentation and the old image used there
removed including translated versions.

The shortcuts section of the Interface preferences was revised to reflect
the above, and to remove the Show menu mnemonics text because that option
is not available anymore.

Closes #259

(cherry picked from commit 221ca791cef7b4697970163eb0ccea719d618a70)