Update links referencing wiki.gimp.org
We had a few remaining links to wiki.gimp.org which is gone.
I replaced one with a link to the new roadmap on the developer website,
others with a WaybackMachine link.
The already disabled entries in the Bibliography were removed.

I completely removed docs/Manualstructure.txt since it didn't give any
extra useful information that can't be seen from the index page of our
documentation. The information in WaybackMachine didn't give any newer
info.