build: CI should fail on validate error
Our normal make validate-all doesn't stop on errors. Since it is such a
length process it is better to continue and later check the logs.

However, for our CI this is not what we want. A job needs to fail on
failed validation. So we add a test after validation whether any xml files
exist inside the log folder, since that is the indication of validation
failure.

We do make po before validation, because I first noticed that the XML files
got built again in the "make" stage later on. After thinking about it, doing
this will also better ensure up-to-date translations in the XML files for
the validation stage.