build: remove foreign from AM_INIT_AUTOMAKE
In the past we added foreign because README.md was not recognized as
README, which is required for the default (gnu) automake build.

However, we now found out this causes a problem, because certain files,
like INSTALL, only get created in the default setting, not when foreign
is defined. This means that make dist fails.

I saw online that automake apparently started allowing README.md as
replacement for README since version 1.16.2. Since even Debian stable
already has a higher version, let's try removing foreign.