build: remove foreign from AM_INIT_AUTOMAKE
In the past we added foreign because README.md was not recognized as
README, which is required for the default (gnu) automake build.

However, we now found out this causes a problem, because certain files,
like INSTALL, only get created in the default setting, not when foreign
is defined. This means that make dist fails.

I saw online that automake apparently started allowing README.md as
replacement for README since version 1.16.4 from 2021. Since Debian testing
uses 1.16.5, let's remove foreign and set automake version to 1.16.

Also add a note to our README.md informing package creators of this
requirement.