Add pl (Polish) language to LINGUAS
So far, Polish was only enabled for QUICKREFERENCE_ALL_LINGUAS. Since we
intend to simplify our language updating procedures, and Polish was the
only language only available in quickreference, this setting will be
removed.

As discussed in issue #318 we are adding empty Polish po files for the
other parts of the GIMP manual.