Issue #368: renamed layer crop menu commands
Crop to Selection and Crop to Content in the Layer menu were renamed to
Resize Layers to Selection (since it also allows for enlarging the layer),
and Crop Layers to Content.

In addition to that make clear that these commands now work on selected
layers.

We also separated both commands into their own file. This allows them to
both be shown in the file menu list.