Update example images showing incorrect use of use sRGB
Cropped the apple-banana image from the Save Color Profile documentation
page to only show the image.
Used the LargeRGB-elle-V4-g10.icc from Elle Stone's Github to convert it
to that profile and added that xcf to our repository so it can be reused.
Moved the base image "apple-banana.png" to our base-examples folder within
our images folder.
Removed the old example image and replaced it with an updated image that
does not show GIMP's interface to reduce the work for translators.

Closes #389.