ci: fix distcheck-jobs failure due to shorter timeout
When we added the distcheck jobs, the timeout for a job was 2 hours.
It seems they changed this back to 1 hour, like it was in the past,
which causes the distcheck jobs to fail, due to needing more than
1 hour to finish.

Let's split these jobs the same way we do with build-debian jobs:
all languages split over 4 jobs instead of 2.