Navigation Dialog: Replace UI screenshot with corresponding icon
Replace custom screenshot which stressed the navigation icon/button in the UI with the corresponding (legacy) icon inline. That's what the rest of the text also does/uses, and it's a small step to allow us easier "icon theme maintenance" in the future.

Also make the text clearer by mentioning the scrollbars, and cover that quick access to this navigation button is only available when scrollbars are enabled.