Use consistent filename scheme for Tool Options screenshots
As a documentation maintainer, I'd like to be able to quickly check a related set of screenshots without having to grep Docbook pages to identify image filenames.

As a translator, I'd like to be able to prioritize my work and quickly find the most important sets of screenshots to localize.

As both, I'd like to be able to find docs related to tools by the tool's name instead of cryptic concatenations of abbreviations.