Remove "fairly recent change"
Such phrases accumulate technical debt as they are hard to ever find again for documentation writers when not explicitly mentioning a specific software version to grep for.

As a user, when was "fairly recently" and how is that relevant for me using (some certain version of) the software right now while not necessarily knowing or caring about the software release history?