Use a consistent naming scheme for cursor screenshots
Introduce a shared filename prefix for cursor screenshots which allows mapping screenshots to GIMP's internal code variable names.

As a documentation writer, I'd like to be able to find (and potentially update) all screenshots of GIMP cursors without having to grep Docbook pages and get incomplete results (due to not consistently using the string "cursor" anyway).

As a documentation writer, I'd like to be able to realize that a screenshot depicts a custom GIMP cursor (and not a stock icon to unsuccessfully search for) when it comes to documentation maintenance.

Thus stop using random, non-self-explanatory filenames.