Convert Mypaint brush Hardness example images into separate files
Don't force translators to edit an XCF bitmap image file to localize text.

Instead, make each example an image file on its own which includes no translatable text, and move the text into the text description of each figure (image) so it can be localized via editing the corresponding .po file.

Also drop the information that the example brush does not act below 0.70 as it's irrelevant for the reader to understand the example images.