Fix invalid XML in Italian due to help-id changes
When help-ids in a translation are used that are not used in the
original, then changes to those help-ids are not detected when updating
po files. This was the case here.

We should probably add some validation that detects these cases to
our CI. Translations should follow the original text and not add
things that are not in the original. The right thing to do in these
cases would be to open an issue to improve the original text.