tools: fix xml2po to be able to use parameterized external entities
The translated XML files were missing the parameterized external
entities: `%entities;` so we hack the output to add that back in.

I wasn't able to figure out how to make the serialization keep this,
so we parse the external entities and save a corrected copy.
Then at the end for the complete output we replace the entities
(if any) with the corrected version.