basic concepts: improve text a bit and add some links
Add an actual link where Wilber's construction kit can be downloaded
instead of pointing at the wrong location in the source that isn't
available for everyone.
Not that this is very useful. It is low resolution and doesn't reflect
the current Wilber version anymore. We should probably think about
removing this or pointing to an updated Wilber.