ci: Use Debian unstable to fixing our CI
The libxml2 version in Debian stable and testing is causing our CI to
fail. This bug has been fixed in newer versions of libxml2, but there
are issues updating it for Debian due to API breakage.

Tests show that with a newer version of libxml2 used in Debian unstable,
our ci does succeed.
To be able to move forward, let's move to Debian unstable for now,
but we should probably go back to Debian testing as soon as possible.

Fixes #460 with respect to our CI. I will do a separate commit in
gimp-web to fix the online help generation.