Update Dialogs button bar screenshot
"Button Bar on the Brushes dialog" has six buttons, not five.
The accompanying text makes it sufficiently clear to find this "on
the bottom of the dialog" so it's fine not to show the full dialog.

Also remove localized screenshots as the new screenshot has no
localizable content anymore, thus less maintenance costs.