tools: add a list of help-ids to ignore
Certain help ids are not liked to any dialog in GIMP and thus won't ever
be called as context sensitive help from within GIMP.
There is no need to list these as missing help ids in our documentation.

From within show-translation_progress.py we check if an id appears in
this list and if it is, it won't be listed as a missing help-id.

I did a quick browse over the plug-ins to fill the ignore list, but it is
possible I missed or misinterpreted something.