Separate translatable and non-translatable example screenshot content
As a translator, I want to be able to provide localized screenshots
of UI elements such as dialogs, without having to acquire knowledge
how to perform image manipulation.

After preparing the application, I simply want to take a screenshot
of whatever is supposed to be depicted, without having to fiddle with
additional non-localizable example content in the same image file.