Fix #519 ci build failure on Debian unstable
This was caused by a newer version of automake than we were testing for.

Increase highest minor version to 17 and adjust the error message
to make it easier to figure out that we may need to increase this
version again in the future.

(cherry picked from commit 8ea580e44c77c2d61a14cb114809092f51646f04)