Update Zoom tool documentation
As mentioned in https://gitlab.gnome.org/GNOME/gimp/-/issues/12939
the Auto-resize window setting is only meant for MWM. In addition,
it doesn't seem to work at all at the moment.
Also make some small adjustments to the overall text and add a help-id
to be able to directly link to the zoom preferences.