GIMP renamed/removed a few path related help-ids
For 3.1.4 and 3.2 `gimp-view-snap-to-vectors` was renamed to
`gimp-view-snap-to-path` and `gimp-tool-vectors` was replaced by
`gimp-tool-path` which already existed and directly links to the
Path tool.

To be able to use this also in our 3.0 documentation we keep the old
ids available als alternate anchors. That way we don't have to wait
with committing until we branch of 3.0.